changed datatype

+++ src/dataset.py
@@ -155,11 +155,11 @@ class VectorGetter():
 
 
         y = self.y_label_encoder.transform(y).reshape(-1, 1)
-        y = self.y_onehot_encoder.transform(y).todense().astype(np.int32)
+        y = self.y_onehot_encoder.transform(y).todense().astype(np.byte)
 
 
 
-        X = np.array(X, dtype=np.int32)
+        X = np.array(X, dtype=np.byte)
         # print(len(y), "individual tracks loaded!")
 
         return X, y
@@ -193,8 +193,8 @@ class VectorGetter():
 
 
         y = self.y_label_encoder.transform(y).reshape(-1, 1)
-        y = self.y_onehot_encoder.transform(y).todense().astype(np.int32)
-        X = np.array(X, dtype=np.int32)
+        y = self.y_onehot_encoder.transform(y).todense().astype(np.byte)
+        X = np.array(X, dtype=np.byte)
 
         X_train, X_test, y_train, y_test = train_test_split(X, y)
